page's text's translation information
The website has text on it, for buttons and else. In order to make sure that website is capable of being used by English, Russian and Georgian speakers comfortably, there must be a translation happening to switch the website's language from one language to the next. That is where the translations for the text comes in.

The newly added data.json file is a JSON array file containing the text for the website in English, Russian and Georgian, as well as the IDs of the objects that contain the texts.