Added list of pages
In order to keep everything organized I have chosen to use a file which has the names of all the would be pages of the website, this it to be used as a way of ensuring that the website runs cleanly.

NOTE: I have changed the name of about_keto_info.txt file to data.txt file. Located at /data/about_keto/ folder. Just a minor change but worthy of a mention.